add review and finish note updating, and fix typo + finish stripe

+++ pcs_api/src/Entity/Appartement.php
@@ -175,12 +175,12 @@ class Appartement
         return $this;
     }
 
-    public function getNote(): ?int
+    public function getNote(): ?float
     {
         return $this->note;
     }
 
-    public function setNote(int $note): static
+    public function setNote(float $note): static
     {
         $this->note = $note;
